+++ src/mir_lower.rs
@@ -966,6 +966,11 @@ impl<'c> FnLowerer<'c> {
                 self.b.assign(dst, mir::RValue::Load(id));
                 Ok(())
             }
+            hir::RValue::MoveOut(op) => {
+                let id = self.lower_operand(op)?;
+                self.b.assign(dst, mir::RValue::MoveOut(id));
+                Ok(())
+            }
             hir::RValue::Panic(msg) => {
                 self.b.assign(dst, mir::RValue::Panic(msg.clone()));
                 Ok(())
@@ -2605,7 +2610,9 @@ fn rvalue_use_locals(op: &mir::RValue, out: &mut Vec<mir::LocalId>) {
         | StateGet { .. } => {}
         mir::RValue::Resume(x) => out.push(*x),
         ReceiveWait { timeout, .. } => out.push(*timeout),
-        Load(x) | ArrayLen(x) | Unary(_, x) | CapabilityCheck { val: x } => out.push(*x),
+        Load(x) | MoveOut(x) | ArrayLen(x) | Unary(_, x) | CapabilityCheck { val: x } => {
+            out.push(*x)
+        }
         PerformAsync { args, .. } => out.extend(args.iter().copied()),
         LoadFieldNamed { obj, .. } | LoadFieldPos { obj, .. } => out.push(*obj),
         ArrayLoad { arr, idx } => {
@@ -2710,7 +2717,7 @@ fn walk_hir_operand(op: &hir::Operand, acc: &mut HashSet<String>) {
 
 fn walk_hir_rvalue(rv: &hir::RValue, acc: &mut HashSet<String>) {
     match rv {
-        hir::RValue::Use(op) => walk_hir_operand(op, acc),
+        hir::RValue::Use(op) | hir::RValue::MoveOut(op) => walk_hir_operand(op, acc),
         hir::RValue::Literal(_, _) | hir::RValue::SelfRef(_) | hir::RValue::Panic(_) => {}
         hir::RValue::Block(body) => walk_hir_body(body, acc),
         hir::RValue::Binary(_, l, r, _) => {